Attempt to make lexer more strict.
Characters which are recognized by the lexer as part of a token type
(e.g. curly braces for blocks, colons for symbols) are now forbidden
from existing in other token types (specifically numbers and
symbols/calls).

This makes it impossible to construct, for example, a function call
with a comment in it; this is closing in on what one would expect.